+++ database/migrations/2025_08_11_104426_create_poll_options_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('poll_options')) {
+            return;
+        }
         Schema::create('poll_options', function (Blueprint $table) {
             $table->id();
             $table->foreignId('poll_id')->constrained('post_polls')->onDelete('cascade');
